+++ db/db_utils.py
@@ -1,0 +1,1 @@
+# for getting information from the database
