+++ src/screens/clues.py
@@ -1,0 +1,7 @@
+class Clue:
+    def __init__(self, display, game_state_manager):
+        self.display = display
+        self.game_state_manager = game_state_manager
+
+    def run(self):
+        pass
